Fix textures `relative_path` being incorrectly relative to a non-dir
Blockbench may store the relative paths to model textures in its native
.bbmodel project format when used as a standalone application. However,
this path is relative to the project file instead of the project file
directory, which generates invalid relative paths, because such paths
are only meant to be relative to directories and not files.

Fix this problem by calling `dirname` on the project save path that's
used to relativize texture paths. The resulting relative paths have been
tested to work well across platforms.