fix the inconsistency of using UV Overlay
The UV Overlay is now renamed to UV Show All, as now it's only function is to show the UV from all objects.
     (it was its only function before, but it was confusing to use, so I made it more clear what it does).
The other UV options are shown normally when clicking with the Right Mouse Button in the UVEditor.
The settings option regarding the UV Mode is now an Enum, to properly represent what's happening