Update pt_br.json
Updated and corrected code. Before creating this code with translations to Brazilian Portuguese, I had already entered the site https://poeditor.com/join/project/EFP1ygSsn7, but it has only the European Portuguese and even so would be a lot of work to correct and adapt a text from another branch of Portuguese. 

It would be interesting to create one for Brazil, or use this code that is already being praised by the Brazilian community.